Organize files into corresponding folders
