+++ client/src/components/Errors/NotFound.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import '../notFound.scss';
+import '../../notFound.scss';
 
 const NotFound = () => {
   return (
