+++ client/src/components/HotTakes/HotTakeList.js
